Move resolver
